+++ src/day_04/puzzle_02.rs
@@ -1,4 +1,4 @@
-pub fn solve(content: &str) -> i32 {
+pub fn solve(content: &str) -> i128 {
     let mut count = 0;
 
     let letters: Vec<Vec<char>> = content.lines().map(|l| l.chars().collect()).collect();
